+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Mongo DB and Mongoose
 - Adds Redis
 - Adds Bull for async + scheduled Jobs
+- Add example conversation